Add support for retrieving compositions from API
- Add new DataPersistanceType to specify source of data from repositories
- Update Project model, add dataPersistanceType field and getter and setter
- Update ProjectRepository.getById to get composition details from API in cascade and merge with local result by greater last modification date
- Add ProjectRepository.returnLastModified helper method to ProjectRepository to return last modified project
- Update ProjectRepository.getListProjectsByLastModificationDescending to merge results from realm and api data sources
- Add ProjectRepository.mergeCompositions helper method to merge project lists from both data sources
- Fix ProjectRepository updateFrameRate and setWatermarkActivated implementations that where doing a recursive call to themselves
- Add CompositionApiDataSource.getById implementation
- Update CompositionToCompositionDtoMapper constructor to set new fields rootPath and privatePath values
- Add CompositionToCompositionDtoMapper.reverseMap implementation
- Add TrackToTrackDtoMapper.reverseMap implementation
- Update DeleteProjectUseCase.delete with a workarround to prevent a NPE when project has audioTracks empty
- Update GalleryProjectListActivity.showLoading to runOnUiThread
- Update GalleryProjectListAdapter.onBindViewHolder to show cloud icon on project thumbnail if project is only API backed
- Update DateUtils to support API date format; update local date format for 24h and timezon support; add parseStringDate to return a Date object with parsed string value from project
- Add CompositionApiClient.get method to get composition details from API in cascade
- Add CompositionService.get route; fix updateComposition route compositionId @Path param
- Update CompositionDto, add tracks getter; add getProductTypeList to parse productType field into a list
- Update CompositionQuery to support cascade parameter
- Update TrackDto, rename trackId to trackIndex, mute to muted, fix compositionId type to String; add constructor for initializing mediaItems list; add mediaItems getter
- Update gallery_project_view_holder layout to include new ic_cloud_queue_white_24dp icon

+++ app/src/main/java/com/videonasocialmedia/vimojo/galleryprojects/domain/DeleteProjectUseCase.java
@@ -4,6 +4,7 @@ import com.videonasocialmedia.videonamediaframework.model.Constants;
 import com.videonasocialmedia.videonamediaframework.model.media.Media;
 import com.videonasocialmedia.videonamediaframework.model.media.Music;
 import com.videonasocialmedia.videonamediaframework.model.media.Video;
+import com.videonasocialmedia.videonamediaframework.model.media.track.AudioTrack;
 import com.videonasocialmedia.videonamediaframework.model.media.track.Track;
 import com.videonasocialmedia.vimojo.composition.domain.model.Project;
 import com.videonasocialmedia.vimojo.repository.music.MusicDataSource;
@@ -13,6 +14,7 @@ import com.videonasocialmedia.vimojo.asset.repository.datasource.VideoDataSource
 import com.videonasocialmedia.vimojo.utils.FileUtils;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.LinkedList;
 
 import javax.inject.Inject;
@@ -22,22 +24,21 @@ import javax.inject.Inject;
  */
 
 public class DeleteProjectUseCase {
-
   protected ProjectRepository projectRepository;
   protected VideoDataSource videoRepository;
-  protected MusicDataSource musicRepository;
-  protected TrackDataSource trackRepository;
+  private MusicDataSource musicRepository;
+  private TrackDataSource trackRepository;
 
   @Inject
   public DeleteProjectUseCase(ProjectRepository projectRepository, VideoDataSource videoRepository,
-                              MusicDataSource musicRepository, TrackDataSource trackRepository){
+                              MusicDataSource musicRepository, TrackDataSource trackRepository) {
     this.projectRepository = projectRepository;
     this.videoRepository = videoRepository;
     this.musicRepository = musicRepository;
     this.trackRepository = trackRepository;
   }
 
-  public void delete(Project project){
+  public void delete(Project project) {
     Track videoTrack = project.getMediaTrack();
     LinkedList<Media> videoList = videoTrack.getItems();
     for (Media media : videoList) {
@@ -45,14 +46,18 @@ public class DeleteProjectUseCase {
     }
     trackRepository.remove(videoTrack);
 
-    Track musicTrack = project.getAudioTracks().get(Constants.INDEX_AUDIO_TRACK_MUSIC);
-    LinkedList<Media> musicList = musicTrack.getItems();
-    for(Media media: musicList){
-      musicRepository.remove((Music) media);
+    ArrayList<AudioTrack> audioTracks = project.getAudioTracks();
+    if (audioTracks != null && audioTracks.size() > 0) {
+      // TODO(jliarte): 9/08/18 FIXME: sometimes audioTracks is empty
+      Track musicTrack = audioTracks.get(Constants.INDEX_AUDIO_TRACK_MUSIC);
+      LinkedList<Media> musicList = musicTrack.getItems();
+      for (Media media: musicList){
+        musicRepository.remove((Music) media);
+      }
+      trackRepository.remove(musicTrack);
     }
-    trackRepository.remove(musicTrack);
 
-    if(project.hasVoiceOver()) {
+    if (project.hasVoiceOver()) {
       Track voiceOverTrack = project.getAudioTracks().get(Constants.INDEX_AUDIO_TRACK_VOICE_OVER);
       LinkedList<Media> voiceOverList = voiceOverTrack.getItems();
       for (Media media : voiceOverList) {
